refactor: move BaseComponent into components folder
