+++ src/routes/landing/About.svelte
@@ -1,6 +1,6 @@
 <script>
 	import timeline from '$lib/images/timeline.svg';
-	import BaseComponent from './BaseComponent.svelte';
+	import BaseComponent from '../components/BaseComponent.svelte';
 </script>
 
 <BaseComponent headline={'Über uns'} id={'about'}>
